Create Solution.cpp
The first and last element of the array is adjacent. So, loop from the second element till n-1. Take each vertices and form a triangle with it and the first and last element.

In top-down approach, since we move ' i ' front and ' j ' toward back , so in bottom - up we will take i back and j to front and rest all the stuffs are same. (Note -: The bottom - up approach is always just opposite of the Top-down approach)